Extract calculateWrappedText() from renderer
